# openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Industrial Lathe Tools API",
    "description": "Industrial Lathe Tools API - Complete product management system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/hesabfa/status": {
      "get": {
        "tags": [
          "Hesabfa"
        ],
        "summary": "Hesabfa integration status",
        "operationId": "hesabfa_status_api_v1_hesabfa_status_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "karzar_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Karzar Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HesabfaStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/hesabfa/items/push": {
      "post": {
        "tags": [
          "Hesabfa"
        ],
        "summary": "Push site products into Hesabfa as item shells (qty 0)",
        "operationId": "push_items_api_v1_hesabfa_items_push_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "karzar_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Karzar Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HesabfaItemPushResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hesabfa/stock/sync": {
      "post": {
        "tags": [
          "Hesabfa"
        ],
        "summary": "Deprecated: Hesabfa→site quantity pull is disabled",
        "operationId": "sync_stock_api_v1_hesabfa_stock_sync_post",
        "deprecated": true,
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "karzar_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Karzar Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HesabfaStockSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hesabfa/sales-summary": {
      "get": {
        "tags": [
          "Hesabfa"
        ],
        "summary": "Website-only paid sales (Hesabfa admin reads disabled)",
        "description": "Website paid totals only — never returns Hesabfa invoice/sales figures.",
        "operationId": "sales_summary_api_v1_hesabfa_sales_summary_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "karzar_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Karzar Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HesabfaSalesSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HesabfaItemPushResponse": {
        "properties": {
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "errors": {
            "type": "integer",
            "title": "Errors"
          },
          "error_samples": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Error Samples"
          }
        },
        "type": "object",
        "required": [
          "created",
          "updated",
          "skipped",
          "errors"
        ],
        "title": "HesabfaItemPushResponse"
      },
...
      "HesabfaSalesSummaryResponse": {
        "properties": {
          "website_paid_total_toman": {
            "type": "string",
            "title": "Website Paid Total Toman",
            "description": "Sum of paid purchase orders on the site (Tomans)"
          },
          "website_paid_order_count": {
            "type": "integer",
            "title": "Website Paid Order Count"
          },
          "hesabfa_sales_total": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hesabfa Sales Total",
            "description": "Deprecated: always null (admin Hesabfa reads disabled)"
          },
          "hesabfa_sales_total_toman": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hesabfa Sales Total Toman",
            "description": "Deprecated: always null (admin Hesabfa reads disabled)"
          },
          "hesabfa_invoice_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hesabfa Invoice Count"
          },
          "hesabfa_currency_unit": {
            "type": "string",
            "title": "Hesabfa Currency Unit"
          },
          "hesabfa_available": {
            "type": "boolean",
            "title": "Hesabfa Available",
            "default": false
          },
          "hesabfa_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hesabfa Error",
            "default": "hesabfa_admin_reads_disabled"
          }
        },
        "type": "object",
        "required": [
          "website_paid_total_toman",
          "website_paid_order_count",
          "hesabfa_currency_unit"
        ],
        "title": "HesabfaSalesSummaryResponse"
      },
      "HesabfaStatusResponse": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "test_mode": {
            "type": "boolean",
            "title": "Test Mode"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "warehouse_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warehouse Code"
          },
          "currency_unit": {
            "type": "string",
            "title": "Currency Unit"
          },
          "stock_sync_interval_seconds": {
            "type": "integer",
            "title": "Stock Sync Interval Seconds"
          },
          "stock_pull_enabled": {
            "type": "boolean",
            "title": "Stock Pull Enabled",
            "default": false
          },
          "item_push_enabled": {
            "type": "boolean",
            "title": "Item Push Enabled",
            "default": true
          },
          "admin_reads_enabled": {
            "type": "boolean",
            "title": "Admin Reads Enabled",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "enabled",
          "configured",
          "test_mode",
          "base_url",
          "currency_unit",
          "stock_sync_interval_seconds"
        ],
        "title": "HesabfaStatusResponse"
      },
      "HesabfaStockSyncResponse": {
        "properties": {
          "checked": {
            "type": "integer",
            "title": "Checked"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "unchanged": {
            "type": "integer",
            "title": "Unchanged"
          },
          "missing_in_hesabfa": {
            "type": "integer",
            "title": "Missing In Hesabfa"
          },
          "disabled": {
            "type": "boolean",
            "title": "Disabled",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "checked",
          "updated",
          "unchanged",
          "missing_in_hesabfa"
        ],
        "title": "HesabfaStockSyncResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}